Partial refacactor of existing code into service module
