+++ codegen-service/llm-generation-library/src/main/java/org/rj/modelgen/llm/Main.java
@@ -1,0 +1,7 @@
+package org.rj.modelgen.llm;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
