--- codegen-service/llm-generation-service/src/main/java/org/rj/codegen/codegenservice/config/Config.java
@@ -1,10 +1,0 @@
-package org.rj.codegen.codegenservice.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
-
-@Configuration
-class Config { }
