--- codegen-service/llm-generation-service/src/main/java/org/rj/codegen/codegenservice/gpt/beans/ExecutionContext.java
@@ -1,7 +1,0 @@
-package org.rj.codegen.codegenservice.gpt.beans;
-
-public enum ExecutionContext {
-    None,
-    Groovy,
-    BPMN
-}
