--- codegen-service/llm-generation-service/src/main/java/org/rj/codegen/codegenservice/gpt/beans/SubmissionResponse.java
@@ -1,126 +1,0 @@
-package org.rj.codegen.codegenservice.gpt.beans;
-
-import java.util.List;
-
-public class SubmissionResponse {
-    private String id;
-    private String object;
-    private Long created;
-    private String model;
-    private Usage usage;
-    private List<Choice> choices;
-
-    public SubmissionResponse() { }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getObject() {
-        return object;
-    }
-
-    public void setObject(String object) {
-        this.object = object;
-    }
-
-    public Long getCreated() {
-        return created;
-    }
-
-    public void setCreated(Long created) {
-        this.created = created;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
-    public void setModel(String model) {
-        this.model = model;
-    }
-
-    public Usage getUsage() {
-        return usage;
-    }
-
-    public void setUsage(Usage usage) {
-        this.usage = usage;
-    }
-
-    public List<Choice> getChoices() {
-        return choices;
-    }
-
-    public void setChoices(List<Choice> choices) {
-        this.choices = choices;
-    }
-
-    public static class Usage {
-        private Integer prompt_tokens;
-        private Integer completion_tokens;
-        private Integer total_tokens;
-
-        public Usage() { }
-
-        public Integer getPrompt_tokens() {
-            return prompt_tokens;
-        }
-
-        public void setPrompt_tokens(Integer prompt_tokens) {
-            this.prompt_tokens = prompt_tokens;
-        }
-
-        public Integer getCompletion_tokens() {
-            return completion_tokens;
-        }
-
-        public void setCompletion_tokens(Integer completion_tokens) {
-            this.completion_tokens = completion_tokens;
-        }
-
-        public Integer getTotal_tokens() {
-            return total_tokens;
-        }
-
-        public void setTotal_tokens(Integer total_tokens) {
-            this.total_tokens = total_tokens;
-        }
-    }
-
-    public static class Choice {
-        private ContextEntry message;
-        private String finish_reason;
-        private Integer index;
-
-        public Choice() { }
-
-        public ContextEntry getMessage() {
-            return message;
-        }
-
-        public void setMessage(ContextEntry message) {
-            this.message = message;
-        }
-
-        public String getFinish_reason() {
-            return finish_reason;
-        }
-
-        public void setFinish_reason(String finish_reason) {
-            this.finish_reason = finish_reason;
-        }
-
-        public Integer getIndex() {
-            return index;
-        }
-
-        public void setIndex(Integer index) {
-            this.index = index;
-        }
-    }
-}
