+++ codegen-service/llm-generation-service/src/main/java/org/rj/modelgen/service/config/Config.java
@@ -1,0 +1,10 @@
+package org.rj.modelgen.service.config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+
+@Configuration
+class Config { }
