+++ codegen-service/llm-generation-service/src/main/java/org/rj/modelgen/service/context/ContextProvider.java
@@ -1,0 +1,33 @@
+package org.rj.modelgen.service.context;
+
+import org.rj.modelgen.service.gpt.beans.ContextEntry;
+import org.rj.modelgen.service.gpt.beans.PromptContextSubmission;
+import org.rj.modelgen.service.gpt.beans.SessionState;
+
+import java.util.List;
+
+public abstract class ContextProvider {
+
+    /** Implemented by subclasses **/
+    public abstract PromptContextSubmission buildBody(SessionState session, String prompt);
+    public abstract PromptContextSubmission buildUndecoratedBody(SessionState session, String prompt);
+    public abstract List<String> validateResponse(String response);
+    public abstract String getValidationFailureRetryPrompt(String responseFailingValidation, List<String> validationErrors);
+    public abstract String sanitizeResponse(String response);
+
+    public abstract String generateTransformedOutput(String response);
+
+    protected List<ContextEntry> newContext(String prompt) {
+        return List.of(
+                ContextEntry.forUser(prompt)
+        );
+    }
+
+    protected List<ContextEntry> continuationContext(SessionState session, String prompt) {
+        return List.of(
+                ContextEntry.forAssistant(session.getLastResponse()),
+                ContextEntry.forUser(prompt)
+        );
+    }
+
+}
