+++ codegen-service/llm-generation-service/src/main/java/org/rj/modelgen/service/gpt/beans/ExecutionContext.java
@@ -1,0 +1,7 @@
+package org.rj.modelgen.service.gpt.beans;
+
+public enum ExecutionContext {
+    None,
+    Groovy,
+    BPMN
+}
