--- codegen-service/llm-generation-service/src/test/java/org/rj/codegen/codegenservice/CodegenServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package org.rj.codegen.codegenservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CodegenServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
