+++ codegen-service/llm-generation-service/src/test/java/org/rj/modelgen/service/CodegenServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package org.rj.modelgen.service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CodegenServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
