--- codegen-service/llm-generation/src/main/java/org/rj/codegen/Main.java
@@ -1,7 +1,0 @@
-package org.rj.codegen;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
